fixed compilation issues

+++ src/main/java/com/github/sebhoss/time/FiscalDate.java
@@ -6,7 +6,7 @@
  */
 package com.github.sebhoss.time;
 
-import org.joda.time.LocalDate;
+import java.time.LocalDate;
 
 /**
  * Represents a single date within a fiscal year.
@@ -16,104 +16,104 @@ public interface FiscalDate {
     /**
      * @return The current fiscal year.
      */
-    int getFiscalYear();
+    long getFiscalYear();
 
     /**
      * @return The current month in the current fiscal year.
      */
-    int getFiscalMonth();
+    long getFiscalMonth();
 
     /**
      * @return The current day in the current fiscal year.
      */
-    int getFiscalDayOfYear();
+    long getFiscalDayOfYear();
 
     /**
      * @return The current week in the current fiscal year.
      */
-    int getFiscalWeekOfWeekyear();
+    long getFiscalWeekOfWeekyear();
 
     /**
      * @return The current fiscal year as calendar year.
      */
-    int getCalendarYear();
+    long getCalendarYear();
 
     /**
      * @return The current month in the current fiscal year as calendar month in a calendar year.
      */
-    int getCalendarMonth();
+    long getCalendarMonth();
 
     /**
      * @return The current day in the current month in the current fiscal year as calendar day in a calendar month in a
      *         calendar year.
      */
-    int getCalendarDayOfMonth();
+    long getCalendarDayOfMonth();
 
     /**
      * @return The current day in the current fiscal year as calendar day in a calendar year.
      */
-    int getCalendarDayOfYear();
+    long getCalendarDayOfYear();
 
     /**
      * @return The current week in the current fiscal year as calendar week in a calendar year.
      */
-    int getCalendarWeekOfWeekyear();
+    long getCalendarWeekOfWeekyear();
 
     /**
      * @param years
      *            The years to add to this fiscal date.
      * @return A new fiscal date representing this fiscal date plus the given years.
      */
-    FiscalDate plusYears(int years);
+    FiscalDate plusYears(long years);
 
     /**
      * @param months
      *            The months to add to this fiscal date.
      * @return A new fiscal date representing this fiscal date plus the given months.
      */
-    FiscalDate plusMonths(int months);
+    FiscalDate plusMonths(long months);
 
     /**
      * @param weeks
      *            The weeks to add to this fiscal date.
      * @return A new fiscal date representing this fiscal date plus the given weeks;
      */
-    FiscalDate plusWeeks(int weeks);
+    FiscalDate plusWeeks(long weeks);
 
     /**
      * @param days
      *            The days to add to this fiscal date.
      * @return A new fiscal date representing this fiscal date plus the given days.
      */
-    FiscalDate plusDays(int days);
+    FiscalDate plusDays(long days);
 
     /**
      * @param years
      *            The years to subtract from this fiscal date.
      * @return A new fiscal date representing this fiscal date minus the given years.
      */
-    FiscalDate minusYears(int years);
+    FiscalDate minusYears(long years);
 
     /**
      * @param months
      *            The months to subtract from this fiscal date.
      * @return A new fiscal date representing this fiscal date minus the given months.
      */
-    FiscalDate minusMonths(int months);
+    FiscalDate minusMonths(long months);
 
     /**
      * @param weeks
      *            The weeks to subtract from this fiscal date.
      * @return A new fiscal date representing this fiscal date minus the given weeks;
      */
-    FiscalDate minusWeeks(int weeks);
+    FiscalDate minusWeeks(long weeks);
 
     /**
      * @param days
      *            The days to subtract from this fiscal date.
      * @return A new fiscal date representing this fiscal date minus the given days.
      */
-    FiscalDate minusDays(int days);
+    FiscalDate minusDays(long days);
 
     /**
      * @return This fiscal date as a calendar date.
